requirements.txt
mcp>=1.0.0
httpx>=0.27.0
openai>=1.30.0
anthropic>=0.28.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
